- TrialDistributor::create() now requires pointer to RunPar_t,   this makes it easier to copy required data. - class TrialDistributor - new data members base_level and top_level   trials are shared between base_level up to top_level-1

+++ RunPar_t.cpp
@@ -260,7 +260,7 @@ void RunPar_t::processArguments(int argc, char *argv[])
     // stopgame
     stopgame = a.GetPar<double>("stopgame", 0.0025);
     // seasontrials
-    seasontrials = a.GetPar<int>("seasontrials", 16384);
+    seasontrials = int( a.GetPar<double>("seasontrials", 16384.0) );
     // trials_sharing
     trials_sharing = a.GetPar<string>("trials_sharing", "equal");
     if (!TrialDistributor::isType(trials_sharing))
